+++ junos/func_resource_bgp.go
@@ -445,7 +445,7 @@ func setBgpOptsBfd(setPrefix string, bfdLivenessDetection []interface{},
 				configSet = append(configSet, setPrefixBfd+"authentication key-chain "+bfdLD["authentication_key_chain"].(string))
 			}
 			if bfdLD["authentication_loose_check"].(bool) {
-				configSet = append(configSet, setPrefixBfd+"authentication loose-check")
+				configSet = append(configSet, setPrefixBfd+authenticationLooseCheck)
 			}
 			if bfdLD["detection_time_threshold"].(int) != 0 {
 				configSet = append(configSet, setPrefixBfd+"detection-time threshold "+
@@ -500,7 +500,7 @@ func readBgpOptsBfd(item string, bfdRead map[string]interface{}) error {
 		bfdRead["authentication_algorithm"] = strings.TrimPrefix(itemTrim, "authentication algorithm ")
 	case strings.HasPrefix(itemTrim, "authentication key-chain "):
 		bfdRead["authentication_key_chain"] = strings.TrimPrefix(itemTrim, "authentication key-chain ")
-	case itemTrim == "authentication loose-check":
+	case itemTrim == authenticationLooseCheck:
 		bfdRead["authentication_loose_check"] = true
 	case strings.HasPrefix(itemTrim, "detection-time threshold "):
 		var err error